removed the code in main.dart, created a .dart file for every screen
