+++ .idea/365-delivery.iml
@@ -17,10 +17,13 @@
     <exclude-output />
     <content url="file://$MODULE_DIR$">
       <sourceFolder url="file://$MODULE_DIR$/gen" isTestSource="false" generated="true" />
+      <excludeFolder url="file://$MODULE_DIR$/delivery_app/.dart_tool" />
       <excludeFolder url="file://$MODULE_DIR$/delivery_app/.pub" />
       <excludeFolder url="file://$MODULE_DIR$/delivery_app/build" />
     </content>
     <orderEntry type="jdk" jdkName="Android API 30 Platform" jdkType="Android SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
+    <orderEntry type="library" name="Dart SDK" level="project" />
+    <orderEntry type="library" name="Dart Packages" level="project" />
   </component>
 </module>
